Moved management from core to root.
Refs: #109
