+++ firenado/management/templates/help/init_command_help.txt
