+++ firenado/management/templates/help/main_command_help.txt
