+++ firenado/management/templates/project/__init__.py.txt
