+++ firenado/management/templates/project/firenado.yml.txt
