+++ firenado/management/templates/project/init_command_help.txt
